Intertwine implicit conversions with local desugar
This frees desugar transformations from worrying about inserting
explicit conversions, and frees the implicit conversion visitor
from worry about creating nodes that need to be desugared.

This commit also converts DesugarEnhancedFor into a local
desugar pass.

+++ tests/isolated/ArrayInit.java
@@ -1,6 +1,6 @@
 public class ArrayInit {
     public static void main(String[] args) {
-        int[] a = new int[] {1, 2, 3, 4};
+        int[] a = new int[] {1, 2, 3, (byte) 4};
         print(a);
 
         int x = 1;
